Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@6948 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ javaproxy/src/ImageDialog.java
@@ -82,7 +82,7 @@ class ImageDialog extends Dialog
             int img_width = image.getWidth(this);
             int img_height = image.getHeight(this);
 
-            imgsy = msgsy - (int)(img_height / 2);
+            imgsy = msgsy - (img_height / 2);
             msgsx += img_width;
             if (imgsy <= 0)
                 return;
